Refactor AppBar
